app/requirements.txt
requests>=2.31, <2.32
python-dotenv>= 1.0, <1.1
fastapi>=0.99, <1
redis-om>=0.1, <0.2
uvicorn>=0.22, <0.23